+++ src/lib.rs
@@ -1,16 +1,31 @@
+#![warn(missing_docs)]
+//! # Godot-GGRS-Wrapper
+//! Godot-GGRS-Wrapper exposes different functions to interact with GGRS inside Godot.
+//! All documentation written is explicitly targeted towards use inside Godot, any functions that are usable in Godot have parameters that start with `(&mut self, _owner: &Node)`.
+//! When interacting with the function in Godot you can ignore these 2 parameters and just use what comes after.
+//! For example the [GodotGGRSP2PSession::add_remote_player()] method would just be used like this in Godot: `p2p.add_remote_player("127.0.0.1:7070")`.
+
 use gdnative::prelude::*;
+pub use godotggrs_p2psession::GodotGGRSP2PSession;
+pub use godotggrs_p2pspectatorsession::GodotGGRSP2PSpectatorSession;
+pub use godotggrs_synctestsession::GodotGGRSSyncTestSession;
 
 mod godotggrs_p2psession;
 mod godotggrs_p2pspectatorsession;
-mod godotggrs_synctest;
-mod helper_functions;
+mod godotggrs_synctestsession;
 
+/// Error message that is printed when there's no GGRS session made.
 pub const ERR_MESSAGE_NO_SESSION_MADE: &str = "No session was made.";
+/// Error message that is printed when there's no callback node specified.
 pub const ERR_MESSAGE_NO_CALLBACK_NODE: &str = "No callback node was specified.";
+/// The name of the Godot callback function that gets called when requesting a state save.
 pub const CALLBACK_FUNC_SAVE_GAME_STATE: &str = "ggrs_save_game_state";
+/// The name of the Godot callback function that gets called when requesting a state load.
 pub const CALLBACK_FUNC_LOAD_GAME_STATE: &str = "ggrs_load_game_state";
+/// The name of the Godot callback function that gets called when requesting to advance the frame.
 pub const CALLBACK_FUNC_ADVANCE_FRAME: &str = "ggrs_advance_frame";
 
+/// Routes all Rust panics to Godot so that any uncaught errors are still visible in Godot.
 pub fn init_panic_hook() {
     // To enable backtrace, you will need the `backtrace` crate to be included in your cargo.toml, or
     // a version of rust where backtrace is included in the standard library (e.g. Rust nightly as of the date of publishing)
@@ -52,9 +67,9 @@ pub fn init_panic_hook() {
 }
 
 fn init(handle: InitHandle) {
-    handle.add_class::<godotggrs_p2psession::GodotGGRSP2PSession>();
-    handle.add_class::<godotggrs_synctest::GodotGGRSSyncTest>();
-    handle.add_class::<godotggrs_p2pspectatorsession::GodotGGRSP2PSpectatorSession>();
+    handle.add_class::<GodotGGRSP2PSession>();
+    handle.add_class::<GodotGGRSSyncTestSession>();
+    handle.add_class::<GodotGGRSP2PSpectatorSession>();
     init_panic_hook()
 }