+++ CHANGELOG.md
@@ -4,6 +4,15 @@ In this document, all remarkable changes are listed. Not mentioned are smaller c
 
 ## Unreleased
 
+## 0.4.0
+
+- Rename struct `GodotGGRSSyncTest` to `GodotGGRSSyncTestSession`.
+- Added `get_current_state()` function to `GodotGGRSP2PSession` which returns the current state as a String.
+- Added `get_current_state()` function to `GodotGGRSP2PSpectatorSession` which returns the current state as a String.
+- Added `get_network_stats(handle)` function to `GodotGGRSP2PSession` which returns the network stats of the handle as a tuple.
+- Added `get_network_stats()` function to `GodotGGRSP2PSpectatorSession` which returns the network stats as a tuple.
+- Rename function `receive_callback_node` to `set_callback_node` in all Session structs.
+
 ## 0.3.1
 
 - Updated GGRS to v0.5.0